Refactor pcapng: move InterfaceDescription to its own file
