+++ src/pcapng.rs
@@ -36,7 +36,6 @@
 use crate::blocks::PcapBlock;
 use crate::endianness::*;
 use crate::error::PcapError;
-use crate::linktype::Linktype;
 use crate::traits::*;
 use crate::utils::*;
 use nom::bytes::streaming::{tag, take};
@@ -49,8 +48,10 @@ use rusticata_macros::{align32, newtype_enum};
 use std::borrow::Cow;
 use std::convert::TryFrom;
 
+mod interface_description;
 mod section_header;
 
+pub use interface_description::*;
 pub use section_header::*;
 
 trait PcapNGBlockParser<'a, En: PcapEndianness, O: 'a> {
@@ -276,75 +277,6 @@ pub fn build_ts_f64(ts_high: u32, ts_low: u32, ts_offset: u64, resolution: u64) 
     let ts_fractional = (ts % resolution) as u32;
     // XXX should we round to closest unit?
     ts_sec as f64 + ((ts_fractional as f64) / (resolution as f64))
-}
-
-/// An Interface Description Block (IDB) is the container for information
-/// describing an interface on which packet data is captured.
-#[derive(Debug)]
-pub struct InterfaceDescriptionBlock<'a> {
-    pub block_type: u32,
-    pub block_len1: u32,
-    pub linktype: Linktype,
-    pub reserved: u16,
-    pub snaplen: u32,
-    pub options: Vec<PcapNGOption<'a>>,
-    pub block_len2: u32,
-    pub if_tsresol: u8,
-    pub if_tsoffset: i64,
-}
-
-impl<'a> InterfaceDescriptionBlock<'a> {
-    /// Decode the interface time resolution, in units per second
-    ///
-    /// Return the resolution, or `None` if the resolution is invalid (for ex. greater than `2^64`)
-    #[inline]
-    pub fn ts_resolution(&self) -> Option<u64> {
-        build_ts_resolution(self.if_tsresol)
-    }
-
-    /// Return the interface timestamp offset
-    #[inline]
-    pub fn ts_offset(&self) -> i64 {
-        self.if_tsoffset
-    }
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, InterfaceDescriptionBlock<'a>>
-    for InterfaceDescriptionBlock<'a>
-{
-    const HDR_SZ: usize = 20;
-    const MAGIC: u32 = IDB_MAGIC;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], InterfaceDescriptionBlock<'a>, E> {
-        // caller function already tested header type(magic) and length
-        // read end of header
-        let (i, linktype) = En::parse_u16(i)?;
-        let (i, reserved) = En::parse_u16(i)?;
-        let (i, snaplen) = En::parse_u32(i)?;
-        // read options
-        let (i, options) = opt_parse_options::<En, E>(i, block_len1 as usize, 20)?;
-        if block_len2 != block_len1 {
-            return Err(Err::Error(E::from_error_kind(i, ErrorKind::Verify)));
-        }
-        let (if_tsresol, if_tsoffset) = if_extract_tsoffset_and_tsresol(&options);
-        let block = InterfaceDescriptionBlock {
-            block_type,
-            block_len1,
-            linktype: Linktype(linktype as i32),
-            reserved,
-            snaplen,
-            options,
-            block_len2,
-            if_tsresol,
-            if_tsoffset,
-        };
-        Ok((i, block))
-    }
 }
 
 /// An Enhanced Packet Block (EPB) is the standard container for storing
@@ -1077,20 +1009,6 @@ fn if_extract_tsoffset_and_tsresol(options: &[PcapNGOption]) -> (u8, i64) {
         }
     }
     (if_tsresol, if_tsoffset)
-}
-
-/// Parse an Interface Packet Block (little-endian)
-pub fn parse_interfacedescriptionblock_le(
-    i: &[u8],
-) -> IResult<&[u8], InterfaceDescriptionBlock, PcapError<&[u8]>> {
-    ng_block_parser::<InterfaceDescriptionBlock, PcapLE, _, _>()(i)
-}
-
-/// Parse an Interface Packet Block (big-endian)
-pub fn parse_interfacedescriptionblock_be(
-    i: &[u8],
-) -> IResult<&[u8], InterfaceDescriptionBlock, PcapError<&[u8]>> {
-    ng_block_parser::<InterfaceDescriptionBlock, PcapBE, _, _>()(i)
 }
 
 /// Parse a Simple Packet Block (little-endian)
